this should work, but it doesn't
