+++ src/store/index.ts
@@ -20,7 +20,6 @@ import { appStateReducer } from '../reducers';
 import { defaultState } from '../reducers/defaultData';
 import { Actions } from '../actions/Actions';
 import { getLegacyAppState } from './legacy';
-import { AsyncActions } from '../actions/asyncActions';
 
 // This is not very nice, but it's initialized at app startup
 export let persistConfig: FelfelePersistConfig;
@@ -69,30 +68,6 @@ export const initStore = async (initCallback: (store: Store<AppState, Actions>) 
         persistor: persistorInner,
         persistConfig,
     };
-};
-
-export const felfeleInitAppActions = (store: Store<AppState, Actions>) => {
-    // tslint:disable-next-line:no-console
-    console.log('initStore: ', store.getState());
-
-    // @ts-ignore
-    store.dispatch(AsyncActions.cleanupContentFilters());
-    store.dispatch(Actions.cleanFeedsFromOwnFeeds(store.getState().ownFeeds.map(feed => feed.feedUrl)));
-    for (const ownFeed of store.getState().ownFeeds) {
-        store.dispatch(Actions.updateOwnFeed({
-            ...ownFeed,
-            isSyncing: false,
-        }));
-    }
-    // @ts-ignore
-    store.dispatch(AsyncActions.cleanUploadingPostState());
-    store.dispatch(Actions.timeTick());
-    // @ts-ignore
-    store.dispatch(AsyncActions.registerBackgroundTasks());
-    // @ts-ignore
-    store.dispatch(AsyncActions.downloadFollowedFeedPosts());
-
-    setInterval(() => store.dispatch(Actions.timeTick()), 60000);
 };
 
 export const getSerializedAppState = async (): Promise<string> => {
